1- Setting or removing a forced decision will re render only their own flags useDecision. 2- Renamed store to notifier
